Updated mmio to operate on octets, not dwords

+++ include/mem/mmio.h
@@ -6,10 +6,10 @@
 void mem_flush_cache();
 void mem_barrier();
 
-void mem_write(uint32_t address, uint32_t val);
-uint32_t mem_read(uint32_t address);
+void mem_write(uintptr_t address, uint32_t val);
+uint32_t mem_read(uintptr_t address);
 
-void mem_set(uint32_t address, uint32_t len, uint32_t val);
-void mem_copy(uint32_t src, uint32_t dst, uint32_t len);
+void mem_set(uintptr_t address, size_t len, uint8_t val);
+void mem_copy(uintptr_t src, uintptr_t dst, size_t len);
 
 #endif
